requirements/test.txt
# potentially to be used for seperate docker test build

-r ./base.txt

django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin
django-webtest==1.9.4
django-extensions==2.1.5  # https://github.com/django-extensions/django-extensions
mixer==6.1.3 # https://mixer.readthedocs.io/en/latest/
model-mommy==1.6.0 #https://model-mommy.readthedocs.io/en/latest/index.html